+++ todoApp/Presentation/SwiftUI/ViewModel/ViewModelProtocol.SUI.swift
@@ -1,0 +1,26 @@
+//
+//  ViewModelProtocol.swift
+//  todoApp
+//
+//  Created by 조영태 on 2022/10/03.
+//
+import SwiftUI
+import Combine
+
+protocol ViewModelObservableObject: ObservableObject {
+  associatedtype Action
+  
+  func action(_ action: Action)
+}
+
+extension SUI {
+    protocol LoadingProtocol: ViewModelObservableObject {
+        var isShowLoadingIndicator: Bool { get }
+    }
+
+    protocol RetryProtocol: ViewModelObservableObject {
+        var retryError: Error? { get }
+        var retryTrigger: PassthroughSubject<Void, Never> { get }
+    }
+
+}
